Feat: Designed skeleton code for lookup_agenda.py and added user experience feature to import_agenda.py to inform user if the import was successful and how many rows were imported.
TODO: Complete ability to lookup by column and value.

Additional Notes: Something to consider while implementing is that searches should be
case-insensitive to provide a good experience. Additionally when searching for a date,
might be useful to inform the user that dates must be in the format of xx/xx/xxxx with
month/day/year.